requirements.txt
astroid==2.2.5
attrs==19.1.0
bcrypt==3.1.6
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.18.2
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
httplib2==0.12.3
idna==2.8
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.16.3
oauthlib==3.0.1
pandas==0.24.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
pymongo==3.8.0
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
scikit-learn==0.21.1
scipy==1.3.0
six==1.12.0
sklearn==0.0
typed-ast==1.3.5
uritemplate==3.0.0
urllib3==1.25.2
Werkzeug==0.15.4
wrapt==1.11.1
